Remove dependence of TrackSort on Stepper
