+++ test/celeritas/track/TrackSort.test.cc
@@ -12,12 +12,17 @@
 
 #include "corecel/data/Collection.hh"
 #include "celeritas/Types.hh"
+#include "celeritas/ext/GeantPhysicsOptions.hh"
 #include "celeritas/global/CoreTrackData.hh"
 #include "celeritas/global/Stepper.hh"
+#include "celeritas/phys/PDGNumber.hh"
+#include "celeritas/phys/ParticleParams.hh"
+#include "celeritas/phys/Primary.hh"
 #include "celeritas/track/TrackInitParams.hh"
 #include "celeritas/track/detail/TrackSortUtils.hh"
 
-#include "../global/Stepper.test.hh"
+#include "../TestEm3Base.hh"
+#include "celeritas_test.hh"
 
 namespace celeritas
 {
@@ -26,12 +31,59 @@ namespace test
 //---------------------------------------------------------------------------//
 // TEST HARNESS
 //---------------------------------------------------------------------------//
+class TestEm3NoMsc : public TestEm3Base
+{
+  public:
+    GeantPhysicsOptions build_geant_options() const override
+    {
+        auto opts = TestEm3Base::build_geant_options();
+        opts.msc = MscModelSelection::none;
+        return opts;
+    }
 
-using celeritas::units::MevEnergy;
+    //! Make 10GeV electrons along +x
+    std::vector<Primary> make_primaries(size_type count) const
+    {
+        Primary p;
+        p.particle_id = this->particle()->find(pdg::electron());
+        CELER_ASSERT(p.particle_id);
+        p.energy = units::MevEnergy{10000};
+        p.track_id = TrackId{0};
+        p.position = {-22, 0, 0};
+        p.direction = {1, 0, 0};
+        p.time = 0;
+
+        std::vector<Primary> result(count, p);
+        for (auto i : range(count))
+        {
+            result[i].event_id = EventId{i};
+        }
+        return result;
+    }
+};
+
+class TrackSortTestBase : virtual public GlobalTestBase
+{
+  public:
+    //! Create a stepper
+    template<MemSpace M>
+    Stepper<M> make_stepper(size_type tracks)
+    {
+        CELER_EXPECT(tracks > 0);
+
+        StepperInput result;
+        result.params = this->core();
+        result.stream_id = StreamId{0};
+        result.num_track_slots = tracks;
+
+        return Stepper<M>{std::move(result)};
+    }
+};
 
 #define TestTrackPartitionEm3Stepper \
     TEST_IF_CELERITAS_GEANT(TestTrackPartitionEm3Stepper)
-class TestTrackPartitionEm3Stepper : public TestEm3NoMsc
+class TestTrackPartitionEm3Stepper : public TestEm3NoMsc,
+                                     public TrackSortTestBase
 {
   protected:
     auto build_init() -> SPConstTrackInit override
@@ -46,7 +98,8 @@ class TestTrackPartitionEm3Stepper : public TestEm3NoMsc
 
 #define TestTrackSortActionIdEm3Stepper \
     TEST_IF_CELERITAS_GEANT(TestTrackSortActionIdEm3Stepper)
-class TestTrackSortActionIdEm3Stepper : public TestEm3NoMsc
+class TestTrackSortActionIdEm3Stepper : public TestEm3NoMsc,
+                                        public TrackSortTestBase
 {
   protected:
     auto build_init() -> SPConstTrackInit override
@@ -65,14 +118,11 @@ class TestTrackSortActionIdEm3Stepper : public TestEm3NoMsc
 
 TEST_F(TestTrackPartitionEm3Stepper, host_is_partitioned)
 {
-    size_type num_primaries = 8;
-    size_type num_tracks = 128;
-
-    Stepper<MemSpace::host> step(this->make_stepper_input(num_tracks));
-
-    // Initialize some primaries and take a step
-    auto primaries = this->make_primaries(num_primaries);
+    // Create stepper and primaries, and take a step
+    auto step = this->make_stepper<MemSpace::host>(128);
+    auto primaries = this->make_primaries(8);
     step(make_span(primaries));
+
     auto check_is_partitioned = [&step] {
         auto span
             = step.state_ref().track_slots[AllItems<TrackSlotId::size_type>{}];
@@ -83,6 +133,7 @@ TEST_F(TestTrackPartitionEm3Stepper, host_is_partitioned)
                 return status[TrackSlotId{track_slot}] == TrackStatus::alive;
             });
     };
+
     // we partition at the start of the step so we need to explictly partition
     // again after a step before checking
     for (auto i = 0; i < 10; ++i)
@@ -105,15 +156,11 @@ TEST_F(TestTrackPartitionEm3Stepper, host_is_partitioned)
 TEST_F(TestTrackPartitionEm3Stepper,
        TEST_IF_CELER_DEVICE(device_is_partitioned))
 {
-    size_type num_primaries = 8;
-    // Num tracks is low enough to hit capacity
-    size_type num_tracks = num_primaries * 800;
-
-    Stepper<MemSpace::device> step(this->make_stepper_input(num_tracks));
-
     // Initialize some primaries and take a step
-    auto primaries = this->make_primaries(num_primaries);
+    auto step = this->make_stepper<MemSpace::device>(6400);
+    auto primaries = this->make_primaries(8);
     step(make_span(primaries));
+
     auto check_is_partitioned = [&step] {
         // copy to host
         auto const& state_ref = step.state_ref();
@@ -152,14 +199,11 @@ TEST_F(TestTrackPartitionEm3Stepper,
 
 TEST_F(TestTrackSortActionIdEm3Stepper, host_is_sorted)
 {
-    size_type num_primaries = 8;
-    size_type num_tracks = 128;
-
-    Stepper<MemSpace::host> step(this->make_stepper_input(num_tracks));
-
     // Initialize some primaries and take a step
-    auto primaries = this->make_primaries(num_primaries);
+    auto step = this->make_stepper<MemSpace::host>(128);
+    auto primaries = this->make_primaries(8);
     step(make_span(primaries));
+
     auto check_is_sorted = [&step] {
         auto& step_limit = step.state_ref().sim.step_limit;
         auto& track_slots = step.state_ref().track_slots;
@@ -195,15 +239,11 @@ TEST_F(TestTrackSortActionIdEm3Stepper, host_is_sorted)
 
 TEST_F(TestTrackSortActionIdEm3Stepper, TEST_IF_CELER_DEVICE(device_is_sorted))
 {
-    size_type num_primaries = 8;
-    // Num tracks is low enough to hit capacity
-    size_type num_tracks = num_primaries * 800;
-
-    Stepper<MemSpace::device> step(this->make_stepper_input(num_tracks));
-
     // Initialize some primaries and take a step
-    auto primaries = this->make_primaries(num_primaries);
+    auto step = this->make_stepper<MemSpace::device>(6400);
+    auto primaries = this->make_primaries(8);
     step(make_span(primaries));
+
     auto check_is_sorted = [&step] {
         // copy to host
         auto const& state_ref = step.state_ref();
@@ -224,6 +264,7 @@ TEST_F(TestTrackSortActionIdEm3Stepper, TEST_IF_CELER_DEVICE(device_is_sorted))
                 << aid_current << " is larger than " << aid_next;
         }
     };
+
     // A step can change the step-limit action, so we need to redo the sorting
     // after taking a step.
     for (auto i = 0; i < 10; ++i)
